script moved to get_data with python path hack
