+++ get_data/get_from_db.py
@@ -2,8 +2,8 @@ from pathlib import Path
 import json
 from tqdm import tqdm
 
-from get_data.utils import es_utils
-from get_data.utils import s3_utils
+from get_data.src import es_utils
+from get_data.src import s3_utils
 from get_data.cluster_param import ES_INDEX, ES_HOST_IP, ES_HOST_PORT
