+++ get_data/run_manual.py
@@ -1,8 +1,13 @@
 import argparse
 import json
+import sys
+import os
+CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
+PARENT_DIR = os.path.dirname(CURRENT_DIR)
+sys.path.append(PARENT_DIR)
 
-from get_data.utils import training_session as ts
-from get_data.utils import label_handler as lh
+from get_data.src import training_session as ts
+from get_data.src import label_handler as lh
 
 
 def get_args(description):
@@ -18,9 +23,9 @@ def get_args(description):
     return parser.parse_args()
 
 
-def collect_data():
+def run_manual():
     """Run the car in manual mode (control with xbox pad), record pictures and create associated labels."""
-    args = get_args(str(collect_data.__doc__))
+    args = get_args(str(run_manual.__doc__))
     if args.session_template:
         print(f'Session template:\n{json.dumps(lh.Label().get_default_session_template(), indent=4)}')
         exit()
@@ -58,4 +63,4 @@ def collect_data():
 
 
 if __name__ == "__main__":
-    collect_data()
+    run_manual()
