+++ get_data/search_and_download.py
@@ -1,4 +1,9 @@
 import argparse
+import sys
+import os
+CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
+PARENT_DIR = os.path.dirname(CURRENT_DIR)
+sys.path.append(PARENT_DIR)
 
 from get_data import get_from_db
