+++ get_data/src/training_session.py
@@ -12,8 +12,8 @@ from picamera import PiCamera
 from picamera.array import PiRGBArray
 # noinspection PyUnresolvedReferences
 
-from get_data.utils import xbox
-from get_data.utils import label_handler
+from get_data.src import xbox
+from get_data.src import label_handler
 from utils.const import SPEED_NORMAL, IMAGE_SIZE, FRAME_RATE, EXPOSURE_MODE, DIRECTION_C, HEAD_DOWN
 from utils import car_mapping
