+++ get_data/update_db.py
@@ -1,6 +1,6 @@
 
-from get_data.utils import s3_utils
-from get_data.utils import es_utils
+from get_data.src import s3_utils
+from get_data.src import es_utils
 from get_data.cluster_param import ES_INDEX, ES_HOST_PORT, ES_HOST_IP
