+++ get_data/upload_data.py
@@ -1,4 +1,9 @@
 import argparse
+import sys
+import os
+CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
+PARENT_DIR = os.path.dirname(CURRENT_DIR)
+sys.path.append(PARENT_DIR)
 
 from get_data import upload_to_db as upload
 from get_data import cluster_param
