+++ test/test_functional_download.py
@@ -4,8 +4,8 @@ from pathlib import Path
 import time
 
 from get_data import upload_to_db as upload
-from get_data.utils import s3_utils
-from get_data.utils import es_utils
+from get_data.src import s3_utils
+from get_data.src import es_utils
 from get_data import cluster_param
 from get_data import get_from_db
