+++ test/test_label.py
@@ -2,7 +2,7 @@ import pytest
 import json
 from pathlib import Path
 
-from get_data.utils import label_handler as lb
+from get_data.src import label_handler as lb
 from utils.path import SESSION_TEMPLATE_NAME
